Refactor GUI into MVC components
